requirements.txt
slackclient
python-dotenv==0.10.3
jira
flask
flask_cors
gunicorn==19.9.0